refactor!: move demo types to api-demo package
